requirements.txt
streamlit==1.40.2
openai==1.55.3
requests==2.32.3
python-dotenv==1.0.1